backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
flask-restful==0.3.10
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.2
joblib==1.2.0
pymongo==4.3.3
python-dotenv==1.0.0
gunicorn==20.1.0
requests==2.28.2
pyjwt==2.6.0
python-dateutil==2.8.2
gevent==22.10.2
matplotlib==3.7.1
seaborn==0.12.2